requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.6
django-colorfield==0.10.1
django-filter==23.3
django-phone-field==1.8.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
idna==3.4
importlib-metadata==6.8.0
loguru==0.7.2
Markdown==3.5
oauthlib==3.2.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
svgwrite==1.4.3
Tree==0.2.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
win32-setctime==1.1.0
zipp==3.17.0
